Enable fuzzy search
Why:

* It is the most useful search in the beginning.

How:

* The index creation is now customized.
* The sequence field has an analyzer that creates n-grams which allows
  for a fuzzy search.
