+++ src/main/resources/application.conf
@@ -6,4 +6,13 @@ elasticsearch {
   protocol = "http"
   host = "localhost"
   port = 9200
+
+  index {
+    fasta {
+      name = "fasta"
+      minNGram = 3
+      maxNGram = 5
+    }
+
+  }
 }
