+++ src/main/scala/io/beagle/Env.scala
@@ -1,7 +1,10 @@
 package io.beagle
 
+import akka.http.scaladsl.server.Route
 import cats.data.Reader
-import io.beagle.components.{AkkaComponent, ControllerComponent, SettingsComponent}
+import com.typesafe.config.ConfigFactory
+import io.beagle.components.{AkkaComponent, ControllerComponent, Controllers, SettingsComponent}
+import io.beagle.directive.{FileUploadController, SearchSequenceController, Static}
 
 trait Env extends SettingsComponent with ControllerComponent with AkkaComponent
 
@@ -17,4 +20,34 @@ object Env {
 
   val materializer = env map (_.materializer)
 
+  object production extends Env {
+
+    env =>
+
+    override def settings: Settings = new Settings {
+
+      private val config = ConfigFactory.load()
+
+      def uiRoot = config.getString("ui.root")
+
+      def elasticSearch: ElasticSearchSettings = new ElasticSearchSettings {
+
+        def protocol = config.getString("elasticsearch.protocol")
+
+        def host = config.getString("elasticsearch.host")
+
+        def port = config.getInt("elasticsearch.port")
+
+      }
+    }
+
+    def controllers = new Controllers {
+      def fileUpload: Route = FileUploadController.route.run(env)
+
+      def search: Route = SearchSequenceController.route.run(env)
+
+      def static: Route = Static.route.run(env)
+    }
+
+  }
 }
